+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - collapsing games list for previous days
 - calculating game difficulty for teams
 - remembering invite code during sign up
+- rendering active bench boost bonus at points page
+- rendering wasted chips
 
 ## [1.1.1] - 2023-11-10
 ### Added